refactor: 팝업 버전 확인 로직 분리
- VersionManager를 src/version.js로 이동하고 module.exports 가드 추가
- 업데이트 링크 인라인 스타일(cssText)을 popup.css .update-link 클래스로 이동
- 버전 비교 테스트를 vm 로드 대신 직접 require로 전환
