+++ src/popup.html
@@ -80,6 +80,7 @@
 
     <script src="data.js"></script>
     <script src="shared.js"></script>
+    <script src="version.js"></script>
     <script src="popup.js"></script>
   </body>
 </html>
